+++ NetKet/Learning/learning.hpp
@@ -18,7 +18,7 @@
 #include <memory>
 
 #include "Hamiltonian/MatrixWrapper/dense_matrix_wrapper.hpp"
-#include "Stepper/stepper.hpp"
+#include "Optimizer/optimizer.hpp"
 #include "ground_state.hpp"
 
 namespace netket {
@@ -40,9 +40,9 @@ public:
 
       Sampler<MachineType> sampler(graph, hamiltonian, machine, pars);
 
-      Stepper stepper(pars);
+      Optimizer optimizer(pars);
 
-      GroundState le(hamiltonian, sampler, stepper, pars);
+      GroundState le(hamiltonian, sampler, optimizer, pars);
     } else if (method_name == "Ed") {
       Graph graph(pars);
