+++ NetKet/Optimizer/ada_delta.hpp
@@ -12,34 +12,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef NETKET_RMSPROP_HPP
-#define NETKET_RMSPROP_HPP
+#ifndef NETKET_ADADELTA_HPP
+#define NETKET_ADADELTA_HPP
 
+#include "abstract_optimizer.hpp"
 #include <Eigen/Core>
 #include <Eigen/Dense>
 #include <cassert>
 #include <cmath>
 #include <complex>
 #include <iostream>
-#include "abstract_stepper.hpp"
 
 namespace netket {
 
-class RMSProp : public AbstractStepper {
+class AdaDelta : public AbstractOptimizer {
   int npar_;
 
-  double eta_;
-  double beta_;
-
-  Eigen::VectorXd st_;
-
+  double rho_;
   double epscut_;
+
+  Eigen::VectorXd Eg2_;
+  Eigen::VectorXd Edx2_;
 
   const std::complex<double> I_;
 
- public:
+public:
   // Json constructor
-  explicit RMSProp(const json &pars) : I_(0, 1) {
+  explicit AdaDelta(const json &pars) : I_(0, 1) {
     npar_ = -1;
 
     from_json(pars);
@@ -47,31 +46,38 @@ class RMSProp : public AbstractStepper {
   }
 
   void PrintParameters() {
-    InfoMessage() << "RMSProp stepper initialized with these parameters :"
+    InfoMessage() << "Adelta optimizer initialized with these parameters :"
                   << std::endl;
-    InfoMessage() << "Learning Rate = " << eta_ << std::endl;
-    InfoMessage() << "Beta = " << beta_ << std::endl;
+    InfoMessage() << "Rho = " << rho_ << std::endl;
     InfoMessage() << "Epscut = " << epscut_ << std::endl;
   }
 
   void Init(const Eigen::VectorXd &pars) override {
     npar_ = pars.size();
-    st_.setZero(npar_);
+    Eg2_.setZero(npar_);
+    Edx2_.setZero(npar_);
   }
 
   void Init(const Eigen::VectorXcd &pars) override {
     npar_ = 2 * pars.size();
-    st_.setZero(npar_);
+    Eg2_.setZero(npar_);
+    Edx2_.setZero(npar_);
   }
 
   void Update(const Eigen::VectorXd &grad, Eigen::VectorXd &pars) override {
     assert(npar_ > 0);
 
-    st_ = beta_ * st_ + (1. - beta_) * grad.cwiseAbs2();
+    Eg2_ = rho_ * Eg2_ + (1. - rho_) * grad.cwiseAbs2();
+
+    Eigen::VectorXd Dx(npar_);
 
     for (int i = 0; i < npar_; i++) {
-      pars(i) -= eta_ * grad(i) / (std::sqrt(st_(i)) + epscut_);
+      Dx(i) = -std::sqrt(Edx2_(i) + epscut_) * grad(i);
+      Dx(i) /= std::sqrt(Eg2_(i) + epscut_);
+      pars(i) += Dx(i);
     }
+
+    Edx2_ = rho_ * Edx2_ + (1. - rho_) * Dx.cwiseAbs2();
   }
 
   void Update(const Eigen::VectorXcd &grad, Eigen::VectorXd &pars) override {
@@ -81,31 +87,44 @@ class RMSProp : public AbstractStepper {
   void Update(const Eigen::VectorXcd &grad, Eigen::VectorXcd &pars) override {
     assert(npar_ == 2 * pars.size());
 
+    Eigen::VectorXd Dx(npar_);
+
     for (int i = 0; i < pars.size(); i++) {
-      st_(2 * i) =
-          beta_ * st_(2 * i) + (1. - beta_) * std::pow(grad(i).real(), 2);
-      st_(2 * i + 1) =
-          beta_ * st_(2 * i + 1) + (1. - beta_) * std::pow(grad(i).imag(), 2);
-      pars(i) -= eta_ * grad(i).real() / (std::sqrt(st_(2 * i)) + epscut_);
-      pars(i) -=
-          eta_ * I_ * grad(i).imag() / (std::sqrt(st_(2 * i + 1)) + epscut_);
+      Eg2_(2 * i) =
+          rho_ * Eg2_(2 * i) + (1. - rho_) * std::pow(grad(i).real(), 2);
+      Eg2_(2 * i + 1) =
+          rho_ * Eg2_(2 * i + 1) + (1. - rho_) * std::pow(grad(i).imag(), 2);
+
+      Dx(2 * i) = -std::sqrt(Edx2_(2 * i) + epscut_) * grad(i).real();
+      Dx(2 * i + 1) = -std::sqrt(Edx2_(2 * i + 1) + epscut_) * grad(i).imag();
+      Dx(2 * i) /= std::sqrt(Eg2_(2 * i) + epscut_);
+      Dx(2 * i + 1) /= std::sqrt(Eg2_(2 * i + 1) + epscut_);
+
+      pars(i) += Dx(2 * i);
+      pars(i) += I_ * Dx(2 * i + 1);
+
+      Edx2_(2 * i) = rho_ * Edx2_(2 * i) + (1. - rho_) * std::pow(Dx(2 * i), 2);
+      Edx2_(2 * i + 1) =
+          rho_ * Edx2_(2 * i + 1) + (1. - rho_) * std::pow(Dx(2 * i + 1), 2);
     }
   }
 
-  void Reset() override { st_ = Eigen::VectorXd::Zero(npar_); }
+  void Reset() override {
+    Eg2_ = Eigen::VectorXd::Zero(npar_);
+    Edx2_ = Eigen::VectorXd::Zero(npar_);
+  }
 
   void from_json(const json &pars) {
     // DEPRECATED (to remove for v2.0.0)
-    std::string section = "Stepper";
+    std::string section = "Optimizer";
     if (!FieldExists(pars, section)) {
       section = "Learning";
     }
-    eta_ = FieldOrDefaultVal(pars[section], "LearningRate", 0.001);
-    beta_ = FieldOrDefaultVal(pars[section], "Beta", 0.9);
+    rho_ = FieldOrDefaultVal(pars[section], "Rho", 0.95);
     epscut_ = FieldOrDefaultVal(pars[section], "Epscut", 1.0e-7);
   }
 };
 
-}  // namespace netket
+} // namespace netket
 
 #endif
