+++ NetKet/netket.hpp
@@ -23,9 +23,9 @@
 #include "Lookup/lookup.hpp"
 #include "Machine/machine.hpp"
 #include "Observable/observable.hpp"
+#include "Optimizer/optimizer.hpp"
 #include "Sampler/sampler.hpp"
 #include "Stats/stats.hpp"
-#include "Stepper/stepper.hpp"
 #include "Utils/all_utils.hpp"
 
 #endif
