+++ Tutorials/CustomGraph/custom_graph.py
@@ -47,9 +47,9 @@ pars['Sampler'] = {
     'Name': 'MetropolisLocal',
 }
 
-# defining the stepper
+# defining the Optimizer
 # here we use the Stochastic Gradient Descent
-pars['Stepper'] = {
+pars['Optimizer'] = {
     'Name': 'Sgd',
     'LearningRate': 0.05,
 }
