+++ Tutorials/CustomHamiltonian/custom_hamiltonian.py
@@ -62,9 +62,9 @@ pars['Sampler'] = {
     'Name': 'MetropolisLocal',
 }
 
-# defining the stepper
+# defining the Optimizer
 # here we use AdaMax
-pars['Stepper'] = {
+pars['Optimizer'] = {
     'Name': 'AdaMax',
 }
