+++ Tutorials/Ising1d/ising1d.py
@@ -44,9 +44,9 @@ pars['Sampler'] = {
     'Name': 'MetropolisLocal',
 }
 
-# defining the stepper
+# defining the Optimizer
 # here we use the Stochastic Gradient Descent
-pars['Stepper'] = {
+pars['Optimizer'] = {
     'Name': 'Sgd',
     'LearningRate': 0.1,
 }
