+++ Tutorials/Observables/sigmax.py
@@ -58,9 +58,9 @@ pars['Observables'] = {
     'Name': 'SigmaX',
 }
 
-# defining the stepper
+# defining the Optimizer
 # here we use the Stochastic Gradient Descent
-pars['Stepper'] = {
+pars['Optimizer'] = {
     'Name': 'Sgd',
     'LearningRate': 0.1,
 }
